.readthedocs.yaml: Upgrade to v2 format and fix build
According to [1] the readthedocs build is currently broken due to the
error:

ERROR: Could not find a version that satisfies the requirement sphinx-copybutton==0.5.1 (from -r requirements.txt (line 4)) (from versions: 0.1, 0.1.1, 0.1.2, 0.1.3, 0.1.4, 0.1.5, 0.1.6, 0.2.0, 0.2.1, 0.2.2, 0.2.3, 0.2.4, 0.2.5, 0.2.6, 0.2.7.dev0, 0.2.7, 0.2.8, 0.2.9, 0.2.10.dev0, 0.2.10, 0.2.11, 0.2.12, 0.3.0, 0.3.1)
ERROR: No matching distribution found for sphinx-copybutton==0.5.1 (from -r requirements.txt (line 4))

This seems to be due to a change on readthedocs side.

So let's try fixing it by upgrading to a more recent OS version. This also
requires to move to the new v2 config file format [2].

We also take the opportunity to use fixed versions in requirements.txt.

[1]: https://readthedocs.org/projects/tango-controls/builds/
[2]: https://docs.readthedocs.io/en/stable/config-file/v2.html#migrating-from-v1